fixing bugs and cgroup

+++ chat-service/internal/domain/service-discovery.go
@@ -3,6 +3,6 @@ package domain
 type ServerMetadata struct {
 	Name 		string 	`json:"name"`
 	URL 		string	`json:"url"`
-	CPU 		float32	`json:"cpu"`
-	Memory 		float32	`json:"memory"`
+	CPU 		float64	`json:"cpu"`
+	Memory 		float64	`json:"memory"`
 }
